Added logColours & prefixes to all logged messages
Previously logging software would be spammed with the characters used to make the console coloured, being able to turn off coloured logging will allow for that console to be easier to read

Now to prevent this, I have added logColours as a Logger option. This allows you to completely turn off logging in colour to the console. I have also added prefixes for each type of logging. CMMD for commands, INFO for info, WARN for warn, EROR for error and DBUG for debug. These will allow you to clearly see what is logged by what in both coloured logging and non-coloured. This should allow for more easily telling the differences between the different types of logging.