Added getUser function to bot object
getUser allows you to get a user object from either a mention, nickname or username.

Allows the user of the lib to get a user object easily using a string containing the name of the user, either a nickname or a username. It performs a fuzzy search and returns the first user to match. Will default to a mention if one is in the passed message object. If no searchArgs are passed or the message is a DM the message author user object will be returned.

This should allow the user to easily get a user object without having to get it from a mention.