Add permission error messages and fixes
If a command cannot be used in a DM or if the user doesnt have the required
permissions for the command, an error message will now be sent to the user. A
bug was also fixed where commands would not work after the first run because
of the commandUsage map erroring.