Fix event and command execution
Event loading was broken due to an egated `if` check.

Command execution was broken due to the permission check only checking
whether `this.permissions` was undefined before accessing its
properties. The issue with this is that it's always defined, it's at
least null. This should instead check for null instead.